5556c62 Moving gh-pages docs to docs/ folder, with script to build them

+++ build_and_deploy_docs.sh
@@ -1,11 +1,21 @@
-echo "This script, if run from master branch, will build docs website and push it to origin/gh-pages. Make PR from there."
+echo "This script builds docs/ and deploys the site to origin/gh-pages."
+echo "It must be run from master branch, or nothing will happen."
+echo ""
+echo "COMMIT YOUR WORK BEFORE RUNNING THIS."
+echo "This will delete *all* uncomitted files."
+read -p "Have you committed (y/n)? " -n 1 -r
+echo    # (optional) move to a new line
+if [[ ! $REPLY =~ ^[Yy]$ ]]
+then
+    exit 1
+fi
 
 CUR_BRANCH=`git rev-parse --abbrev-ref HEAD`
 MSG=`git log --oneline -1`
 
 if [ $CUR_BRANCH='master' ]; then
     jekyll build -s docs
-    git checkout gh-pages
+    git checkout -b gh-pages
     git rm -qr .
     cp -r _site/. .
     rm -r _site
